Rename file, add plugin
