+++ src/query-tools.js
@@ -1,20 +1,34 @@
+/* eslint-disable unicorn/no-fn-reference-in-iterator */
 const debug = require('debug')('Uttori.StorageProvider.JSON.QueryTools');
 const R = require('ramda');
-const { parseQueryToRamda, validateQuery } = require('uttori-utilities');
+const { parseQueryToRamda, validateQuery, fyShuffle } = require('uttori-utilities');
 
+/**
+ * Processes a query string.
+ * @param {String} query - The SQL-like query to parse.
+ * @param {Object[]} objects - An array of object to search within.
+ * @returns {Object[]} Returns an array of all matched documents.
+ * @example
+ * process('SELECT name FROM table WHERE age > 1 ORDER BY RANDOM LIMIT 3', [{ ... }, ...]);
+ * ➜ [{ ... }, ...]
+ */
 const process = (query, objects) => {
   debug('Processing Query:', query);
   // Filter
-  const { where, order, limit } = validateQuery(query);
-  debug('Found where, order, limit:', where, order, limit);
+  const { fields, where, order, limit } = validateQuery(query);
+  debug('Found fields:', fields);
+  debug('Found where:', where);
+  debug('Found order:', order);
+  debug('Found limit:', limit);
   const whereFunctions = parseQueryToRamda(where);
   const filtered = R.filter(whereFunctions)(objects);
+
   // Sort / Order
-  let sorted;
+  let output;
   if (order[0].prop === 'RANDOM') {
-    sorted = R.sort(() => Math.random() - Math.random(), filtered);
+    output = fyShuffle(filtered);
   } else {
-    sorted = R.sortWith(
+    output = R.sortWith(
       order.map((value) => {
         const sorter = value.sort === 'ASC' ? R.ascend : R.descend;
         return sorter(R.prop(value.prop));
@@ -45,7 +59,16 @@ const process = (query, objects) => {
   // R.sortWith([sorter])(docs);
 
   // Limit
-  return R.take(limit, sorted);
+  if (limit > 0) {
+    output = R.take(limit, output);
+  }
+
+  // Select
+  if (!fields.includes('*')) {
+    output = R.lift(R.pickAll(fields))(output);
+  }
+
+  return output;
 };
 
 module.exports = {
